refactor(ooda_loop): split tests_orient.rs (#1266)
tests_orient.rs 413 -> 250 LOC; new tests_orient_extra.rs (~205 LOC).
